requirements.txt
# Discord Bot Dependencies
discord.py>=2.3.0
wavelink>=3.0.0

# Database
motor>=3.3.0
pymongo>=4.5.0

# Audio Processing
PyNaCl>=1.5.0
opuslib>=3.0.1

# HTTP Requests
aiohttp>=3.8.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
asyncio-mqtt>=0.16.0

# Development Tools
black>=23.0.0
ruff>=0.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: Spotify Integration
spotipy>=2.23.0

# Optional: YouTube Integration
yt-dlp>=2023.10.0

# Optional: Lyrics Service
lyricsgenius>=0.12.0

# Optional: Environment Variables
python-decouple>=3.8 